requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9
pydantic>=1.10
python-dotenv>=1.0
